requirements.txt
yt-dlp>=2023.10.0
openai>=1.0.0
tqdm>=4.66.1
python-dotenv>=1.0.0
pytest>=7.4.3
faster-whisper>=0.10.0; platform_system != "Darwin" or platform_machine != "arm64"  # Optional, not available for M1/M2 Macs